[Refactor] Perform global cleanup

+++ FitnessTracker.Domain/Entities/WorkoutExercise.cs
@@ -4,7 +4,7 @@ public class WorkoutExercise : BaseEntity
 {
     public int Sets { get; set; }
     public int Reps { get; set; }
-    public int WeightUsed { get; set; }
+    public decimal WeightUsed { get; set; }
 
     public Guid WorkoutId { get; set; }
     public Workout? Workout { get; set; }
